+++ src/MixModule.hpp
@@ -316,8 +316,16 @@ struct MixBaseWidget : VenomWidget {
   void appendContextMenu(Menu* menu) override {
     MixBaseModule* module = static_cast<MixBaseModule*>(this->module);
 
-    if (module->mutePresent || module->soloPresent || module->sendPresent || module->panPresent)
-      menu->addChild(new MenuSeparator);
+    menu->addChild(new MenuSeparator);
+    menu->addChild(createMenuLabel("Note: rearrange after adding expanders"));
+    menu->addChild(createMenuItem("Add Mix Offset expander (Only 1, must be 1st)", "", [this](){addExpander(modelMixOffset,this);}));
+    menu->addChild(createMenuItem("Add Mix Mute expander (Only 1)", "", [this](){addExpander(modelMixMute,this);}));
+    menu->addChild(createMenuItem("Add Mix Solo expander (Only 1)", "", [this](){addExpander(modelMixSolo,this);}));
+    menu->addChild(createMenuItem("Add Mix Fade expander (only 1, needs Mute or Solo to left)", "", [this](){addExpander(modelMixFade,this);}));
+    menu->addChild(createMenuItem("Add Mix Fade 2 expander (only 1, needs Mute or Solo to left)", "", [this](){addExpander(modelMixFade2,this);}));
+    if (module->stereo)
+      menu->addChild(createMenuItem("Add Mix Pan expander (only 1)", "", [this](){addExpander(modelMixPan,this);}));
+    menu->addChild(createMenuItem("Add Mix Aux Send expander", "", [this](){addExpander(modelMixSend,this);}));
 
     if (module->mutePresent || module->soloPresent || module->sendPresent)
       menu->addChild(createBoolMenuItem("Soft mute/solo", "",